+++ src/fintrist/migrations/migrate_002.py
@@ -1,0 +1,4 @@
+""""""
+
+def upgrade():
+    pass